Inlined AesCmacFunction for Openssl3AesCmac instead of having separate Generator and Verifier as operations
